moved fragmentHandling to utils package
