+++ app/src/main/java/com/holmal/app/holmal/CreateHousehold.java
@@ -8,9 +8,9 @@ import android.support.v7.app.AppCompatActivity;
 import android.widget.EditText;
 import android.widget.RadioGroup;
 import android.widget.Toast;
-import android.widget.ToggleButton;
 
-import butterknife.BindView;
+import com.holmal.app.holmal.utils.FragmentHandling;
+
 import butterknife.ButterKnife;
 import butterknife.OnClick;
